requirements.txt
flask>=3.0
flask-cors>=4.0
apscheduler>=3.10
requests>=2.31
pandas>=2.1
numpy>=1.24
finance-datareader>=0.9
pykrx>=1.0
tqdm>=4.65
openpyxl>=3.1
lxml>=5.0
gunicorn>=21.2
anthropic>=0.40
google-genai>=1.0
duckdb>=1.0
python-dotenv>=1.0
pypdf>=4.0
yfinance>=0.2.36
